+++ src/main/java/com/liveeasystreet/ecovalue/controller/home/HomeController.java
@@ -1,7 +1,11 @@
 package com.liveeasystreet.ecovalue.controller.home;
 
+import com.liveeasystreet.ecovalue.controller.login.SessionConst;
+import com.liveeasystreet.ecovalue.dto.member.MemberSessionDto;
 import org.springframework.stereotype.Controller;
+import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.SessionAttribute;
 
 @Controller
 public class HomeController {
